pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gmc-geiger-mqtt"
version = "0.2.0"
description = "MQTT bridge for GMC Geiger counters"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE.txt"}
authors = [
    {name = "GMC Geiger MQTT Contributors"},
]
keywords = ["geiger", "mqtt", "radiation", "gmc", "home-assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
    "Topic :: System :: Hardware",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "pyserial>=3.5",
    "pyyaml>=6.0",
    "paho-mqtt>=1.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
gmc-geiger-mqtt = "gmc_geiger_mqtt.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/gmc-geiger-mqtt"
Issues = "https://github.com/yourusername/gmc-geiger-mqtt/issues"
Documentation = "https://github.com/yourusername/gmc-geiger-mqtt/blob/main/README.md"

[tool.hatch.build.targets.wheel]
packages = ["src/gmc_geiger_mqtt"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE.txt",
    "/ARCHITECTURE.md",
    "/pyproject.toml",
]
exclude = [
    "/.venv",
    "/__pycache__",
    "*.pyc",
    "/.pytest_cache",
    "/manual_tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/manual_tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py38"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*" = ["B018"]  # useless expression (for pytest)

[tool.ruff.lint.isort]
known-first-party = ["gmc_geiger_mqtt"]
